requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile setup.cfg
#
anyio==3.2.1
    # via httpcore
beautifulsoup4==4.9.3
    # via archive-md-urls (setup.cfg)
certifi==2021.5.30
    # via httpx
h11==0.12.0
    # via httpcore
httpcore==0.13.6
    # via httpx
httpx==0.18.2
    # via archive-md-urls (setup.cfg)
idna==3.2
    # via
    #   anyio
    #   rfc3986
markdown==3.3.4
    # via archive-md-urls (setup.cfg)
python-dateutil==2.8.2
    # via archive-md-urls (setup.cfg)
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.2.1
    # via beautifulsoup4
tenacity==8.0.1
    # via archive-md-urls (setup.cfg)
